refactor(ui): archi pages — découpe monolithes + primitives + cleanup config
- alternants/page.tsx 1329->321 l. (6 sous-composants présentationnels _components/)
- planning/page.tsx 648->495 l. (PersonView/TableView extraites)
- primitives partagées components/dashboard/ (StatCard, FilterToolbar) ; FilterToolbar adoptée dans history
- suppression code mort config (page-old + 3 *-management anciens, 0 importeur)
Extraction structurelle/présentationnelle uniquement — state/fetch/handlers inchangés.
