+++ app/(dashboard)/planning/page.tsx
@@ -13,16 +13,15 @@ import { PlanningToolbar, type PaintMode, type SlotType } from '@/components/pla
 import { PlanningGrid } from '@/components/planning/planning-grid';
 import { PlanningSidebar } from '@/components/planning/planning-sidebar';
 import { MobileDayView } from '@/components/planning/mobile-day-view';
+import { PersonView } from '@/components/planning/person-view';
+import { TableView } from '@/components/planning/table-view';
 import {
   getWeekDates,
   getWeekNumber,
   getWeekKey,
-  formatDate,
 } from '@/lib/db/utils';
 import type { Employee } from '@/lib/db/schema/employees';
 import type { TimeSlot } from '@/lib/db/schema/schedules';
-import { Avatar, AvatarImage, AvatarFallback } from '@/components/ui/avatar';
-import { ScrollArea } from '@/components/ui/scroll-area';
 
 interface EmployeeWithSchedule extends Employee {
   schedule: { [weekKey: string]: { [day: string]: TimeSlot[] } };
@@ -492,157 +491,5 @@ export default function PlanningPage() {
         setSchedules={setSchedules}
       />
     </div>
-  );
-}
-
-// Person view (secondary)
-function PersonView({
-  employees,
-  daysOfWeek,
-  currentWeekDates,
-  getEmployeeScheduleForDay,
-  getTotalHoursForWeek,
-  slotTypeConfig,
-}: {
-  employees: Employee[];
-  daysOfWeek: string[];
-  currentWeekDates: Date[];
-  getEmployeeScheduleForDay: (id: string, day: string) => TimeSlot[];
-  getTotalHoursForWeek: (id: string) => number;
-  slotTypeConfig: Record<string, { label: string; bgColor: string; borderColor: string; textColor: string }>;
-}) {
-  return (
-    <ScrollArea className="h-full">
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-3 p-1">
-        {employees.map((employee) => (
-          <div
-            key={employee.id}
-            className="rounded-xl border bg-background p-3 border-l-4"
-            style={{ borderLeftColor: employee.color || '#8884d8' }}
-          >
-            <div className="flex items-center gap-2.5 mb-2">
-              <Avatar className="h-9 w-9 ring-2 ring-offset-1" style={{ ['--tw-ring-color' as string]: employee.color || '#8884d8' }}>
-                <AvatarImage src={`https://avatar.vercel.sh/${employee.id}.png`} alt={employee.name} />
-                <AvatarFallback className="text-xs font-bold">{employee.name.split(' ').map((n) => n[0]).join('')}</AvatarFallback>
-              </Avatar>
-              <div className="flex-1 min-w-0">
-                <div className="font-semibold text-sm leading-tight">{employee.name}</div>
-                <div className="text-[11px] text-muted-foreground flex items-center gap-1.5">
-                  {employee.role}
-                  <Badge variant="outline" className="text-[9px] px-1 py-0 h-4 font-bold" style={{ borderColor: employee.color || '#8884d8', color: employee.color || '#8884d8' }}>
-                    {getTotalHoursForWeek(employee.id)}h
-                  </Badge>
-                </div>
-              </div>
-            </div>
-            <div className="space-y-1">
-              {daysOfWeek.map((day, i) => {
-                const slots = getEmployeeScheduleForDay(employee.id, day);
-                const filtered = (day === 'samedi' || day === 'dimanche') ? slots.filter((s) => s.type !== 'vacation') : slots;
-                return (
-                  <div key={day} className="flex items-center gap-2 text-xs">
-                    <span className="w-16 font-medium text-[11px] capitalize text-muted-foreground">{day}</span>
-                    {filtered.length === 0 ? (
-                      <span className="text-[10px] italic text-muted-foreground">Repos</span>
-                    ) : (
-                      <div className="flex flex-wrap gap-1">
-                        {filtered.map((slot, idx) => {
-                          const config = slotTypeConfig[slot.type];
-                          return (
-                            <span key={idx} className={`px-1.5 py-0.5 rounded text-[10px] font-bold ${config?.bgColor} ${config?.textColor} border ${config?.borderColor}`}>
-                              {config?.label} {slot.start}-{slot.end}
-                            </span>
-                          );
-                        })}
-                      </div>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-        ))}
-      </div>
-    </ScrollArea>
-  );
-}
-
-// Table view (secondary)
-function TableView({
-  employees,
-  daysOfWeek,
-  currentWeekDates,
-  getEmployeeScheduleForDay,
-  getTotalHoursForWeek,
-  slotTypeConfig,
-}: {
-  employees: Employee[];
-  daysOfWeek: string[];
-  currentWeekDates: Date[];
-  getEmployeeScheduleForDay: (id: string, day: string) => TimeSlot[];
-  getTotalHoursForWeek: (id: string) => number;
-  slotTypeConfig: Record<string, { label: string; bgColor: string; borderColor: string; textColor: string }>;
-}) {
-  const renderCell = (employeeId: string, day: string) => {
-    const slots = getEmployeeScheduleForDay(employeeId, day);
-    const workSlots = slots.filter((s) => s.type === 'work');
-    const otherSlots = slots.filter((s) => s.type !== 'work');
-    if (workSlots.length === 0 && otherSlots.length === 0) {
-      return <span className="text-[10px] italic text-muted-foreground">Repos</span>;
-    }
-    return (
-      <div className="flex flex-col items-center gap-0.5">
-        {workSlots.map((s, i) => (
-          <span key={i} className="text-[10px] font-medium">{s.start}-{s.end}</span>
-        ))}
-        {otherSlots.map((s, i) => {
-          const config = slotTypeConfig[s.type];
-          return (
-            <span key={`o${i}`} className={`px-1 py-0 rounded text-[9px] font-bold ${config?.bgColor} ${config?.textColor} border ${config?.borderColor}`}>
-              {config?.label}
-            </span>
-          );
-        })}
-      </div>
-    );
-  };
-
-  return (
-    <ScrollArea className="h-full">
-      <div className="rounded-lg border bg-background overflow-hidden">
-        <table className="w-full border-collapse text-xs">
-          <thead className="sticky top-0 bg-muted/80 backdrop-blur-sm z-10">
-            <tr className="border-b">
-              <th className="text-left p-2 text-[11px] uppercase tracking-wider font-medium text-muted-foreground sticky left-0 bg-muted/80 backdrop-blur-sm z-20">Employé</th>
-              {daysOfWeek.map((day, i) => (
-                <th key={day} className="text-center p-2 text-[11px] uppercase tracking-wider font-medium text-muted-foreground">
-                  <div>{day.slice(0, 3)}.</div>
-                  <div className="text-[9px] font-normal">{formatDate(currentWeekDates[i])}</div>
-                </th>
-              ))}
-              <th className="text-center p-2 text-[11px] uppercase tracking-wider font-medium text-muted-foreground">Total</th>
-            </tr>
-          </thead>
-          <tbody>
-            {employees.map((employee) => (
-              <tr key={employee.id} className="border-b hover:bg-muted/30 align-middle">
-                <td className="p-2 sticky left-0 bg-background z-10">
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-2.5 h-2.5 rounded-full flex-shrink-0" style={{ backgroundColor: employee.color || '#8884d8' }} />
-                    <span className="font-medium text-xs">{employee.name}</span>
-                  </div>
-                </td>
-                {daysOfWeek.map((day) => (
-                  <td key={day} className="p-2 text-center">{renderCell(employee.id, day)}</td>
-                ))}
-                <td className="p-2 text-center font-bold text-xs" style={{ color: employee.color || '#8884d8' }}>
-                  {getTotalHoursForWeek(employee.id)}h
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </ScrollArea>
   );
 }
